requirements.txt
requests
selenium
beautifulsoup4
pytz
pyyaml
smtplib
